Add like and dislike info
